libterminus: move print of exit_messsage after it is assigned
Problem: pty_die() can now be called with a NULL message, in which case
the existing exit_message is supposed to be used, but the notification
to the client is printed before the assignment, which results in

 [detached: (null)]

being printed to the user's console.

Move the print of the exit_message to after it is assigned, and print
c->exit_message, not the message argument to the function, so we get
a valid message printed no matter how pty_die() is called.

If exit_message happens to be NULL (unlikely) print "unknown reason"
in its place, rather than "(null)".