shell: do not allow instance owner to use guest shell services
Problem: RPCs to shell services are checked against the current
userid and instance owner, but there isn't any reason to allow the
instance owner to utilize shell services for a guest user job.

Drop the check for FLUX_ROLE_OWNER for shell service requests and
only allow the job user to send RPCs to shell services.