libterminus: fix race on pty attach to short-lived processes
Problem: When a short-lived process is run under a terminus pty,
the process may exit before the client has a chance to attach. If
the pty is also being monitored, then any data read on behalf of the
monitor is lost to the attaching client. Furthermore, in some cases
the entire terminus server may exit if the exiting process causes the
last session to close. This can cause a hang on the client side if the
attach RPC is received by the terminus server code during shutdown,
and so there is never an RPC response.

Rework the pty code to add flags to wait for a client to attach and/or
the pty read side to "close" (EIO) before shutting down the pty.

Allow a caller to get notified of when a pty is "complete" -- i.e.
a client has attached, the pty is closed, and any task using the pty
has exited -- to disentangle pty completion from process exit.

Since a pty is now notified of process exit separately from its
completion, rename flux_pty_close(), which used to take the exiting
process status, to flux_pty_destroy(). A new function flux_pty_exited()
is added for callers to set the exit status of a process using the
pty. flux_pty_exited() may or may not immediately call the completion
callback for the pty, depending on the settings for wait-on-client
and wait-on-close.

On the terminus server side, take advantage of the new functionality
to delay removal of sessions until they are actually complete. Since
sessions may be asynchronosly removed by callback, add a reference
count to terminus session objects and ensure to take a reference in
places where the session could be used after it might be destroyed via
the pty completion callback.

Fixes #4077