shell: pty: fix race in pty.interactive for short lived processes
Problem: There is a race when attaching to a job started with
an interactive pty (-o pty.interactive) when the task is short-lived.
The "attach" RPC to the embedded terminus service may get delivered,
but the shell exits before the response can be sent.

Utilize recent improvements to libterminus to prevent the race.
For an interactive pty, set wait-for-client for the underlying pty so
that the pty completion callback is not issued until the client has
has a chance to attach. Also set wait-on-close to ensure all data
is read before terminating the pty session.

Finally, add a completion reference that is only released when the
terminus server becomes "empty", i.e. there are no more active
sessions. This allws the shell to continue running even if all tasks
have exited, giving a pty client time to attach even for short-lived
processes.

Fixes #4083