t2612-job-shell-pty.t: run flux job attach under runpty
Problem: Some invocations of `flux job attach` on jobs with ptys
are not run under runpty. If the tests are run in an environment
without a tty this will result in strange errors and possibly
test failures.

Add $runpty in front of `flux job attach` invocations where it is
missing.