resource: disallow NULL reason with drain_rank()
Problem: drain_rank() is only used internally in the resource module,
but allows a NULL reason to be passed to the function, arbitrarily
substituting the string "unknown". This is inconsistent with the drain
RPC which allows a NULL reason.

Do not allow a NULL reason to be passed to drain_rank(). This will
encourage resource module entities to always provide a reason so
that there are no mysterious drained ranks with "unknown" reason.