resource: do not encode NULL reason as "unknown" in eventlog
Problem: When a drain RPC is received with no reason, the reason
is encoded as "unknown" in the eventlog, but this means that when
the eventlog is replayed after a restart the drain state is not
accurately reconstructed.

Drop the reason from the resource eventlog entry when it is not set,
instead of changing the provided reason to "unkown."

Unfortunately, the eventlog code doesn't seem to support encode
of JSON null, so use two different reslog_post_pack() statements
depending on if reason is set or not.