broker: add tbon.tcp_user_timeout tunable
Problem: when a broker is turned off, generally the connection
is dropped when a published heartbeat message fails to receive
a tcp acknowledgement after a timeout.  This timeout appears to
be around 15-20m, which may not be responsive enough.

Add the tbon.tcp_user_timeout broker attribute and TOML config
key, which allows the TCP_USER_TIMEOUT socket option to be
configured, and overridden on the broker command line.

Since libzmq is managing the socket and the connections, we
actually have to set the ZMQ_TCP_MAXRT socket.  This value
cannot be updated on a live system.

N.B. on el7, the 0MQ packages (zeromq-4.1.4, czmq-3.0.2) are too
old to support the ZMQ_TCP_MAXRT option, so compile the feature
out if unavailable.

Update flux-config-tbon(5) and flux-broker-attributes(7).
Update sharness test.

Fixes #4116