resource: allow unset reason in drain event replay
Problem: The reason field for a drained set of targets is allowed to
be unset (i.e. NULL), but the drain replay event handler expects a
string on decode.

Relax the unpack arguments during drain event replay to allow the
reason field to be missing. The reason will default to NULL, which
indicates the reason is not currently set.