resource: add overwrite option when draining ranks
Problem: Draining an idset always overwrites the drain reason and
timestamp of all targets, which may not always be the desired outcome.

Add an overwrite argument to update_draininfo_rank(). If overwrite
is 0, then avoid updating a rank if the rank is already drained with
an existing reason set.

Update all callers, using overwrite=1 for now since this preserves
existing behavior.