resource: support optional "mode" in drain RPC
Problem: The drain RPC is currently a big hammer in that it always
overwrites the existing drain reason and timestamp for all targets.
The drain RPC should have safe behavior by default, i.e. overwriting
should not be the default. It may also be useful to only drain
and update ranks that are not already drained.

Switch the default for the resource drain RPC to not overwrite by
default. If any of the supplied target ranks are already drained
with an exising reason, then the entire operation will fail.

Add a new optional "mode" field to the drain RPC which can be
set to either "overwrite", in which case the drain reason will be
overwritten for all target ranks, or "update", in which case only
undrained ranks will be updated with the supplied reason.

In order to properly reconstruct the state, post the overwrite flag
to the eventlog.