shell: add --reconnect option
Problem: The job shell can't survive restart of the local broker.

Add a --reconnect job shell option which adds the FLUX_O_RPCTRACK
option to flux_open(3) flags and adds a comms error handler
with flux_comms_error_set(3) to continually try to reconnect when
the local broker is lost.

On reconnect, the shell's service endpoint is re-registered, and
event subscription re-established.

This should allow the job shell to survive a broker restart when
there is no I/O or other Flux message traffic at the time of the
disconnect.