job-exec: do not try to re-send J to multi-user job with no stdin
Problem: When reattaching to a multiuser sdexec job, sdexec always
tries to re-send the input to the IMP (J), even though the stdin fd
is set to -1. This generates EBADF and a fatal error for the job.

Only send input to the IMP if the current stdin fd is valid. This
effectively avoids resend of J for jobs that were already started.