resource: topo: set mode=update in drain_self()
Problem: The resource.drain RPC in resource/topo.c:drain_self()
will fail with EEXIST if the 'self' rank is already drained.

Considering that an existing, owner-applied reason may be set for the
rank in question, use mode=update on the resource.drain RPC request
in drain_self(). This means the RPC will not return an error if the
rank is already drained and an existing drain reason will not be
updated. The drain event will still be entered into the eventlog.