flux-perilog-run: drain ranks with mode=update
Problem: When the job epilog/prolog as executed by flux-perilog-run
fails on a rank, the rank is drained which overwrites any existing
drain reason.

Add mode=update to the resource.drain RPC as used in flux-perilog-run
so that undrained ranks are drained, but already drained ranks do not
have their reason overwritten.