shell: output: capture log messages to output file
Problem: When job stderr is redirected to a file, job shell log
messages are still directed to the output eventlog. This could
result in an empty output file even when the shell encounters a
fatal error. The user would then be required to run `flux job attach`
after the fact to see errors, which is not ideal.

Add a log message handler to the shell output plugin when stderr is
redirected to a file, so that log messages can be written to the file.

Messages generated before the file is succesfully opened still go to
the output eventlog, but a missing output file should be a good hint
for a user to check on a failed job with `flux job attach`.

Fixes #4117