broker: handle network partition
Problem: If there is a network partition and a TBON child is
disconnected due to TCP ACK failure, and the network partition is
later resolved, the child does not recover.

It was assumed that a message from a LOST child would only be messages
that were in flight before the child received a disconnect control message.
However the resolved network partition case also may result in a message
from a LOST child, and in that case, it is not appropriate to silently
drop the messages.

Since the child has already been disconnected from the instance, send
a disconnect control message so that it can restart the broker and rejoin
the instance.

Fixes #4124