broker: simplify error on module abnormal exit
Problem: Inclusion of `strerror (errno)` in the error message logged
when a module exits mod_main() with a nonzero return code is probably
not helpful to administrators perusing the logs. In most cases,
the module will have already printed a more descriptive error before
exiting with error, and the errno at best is not helpful and worst
may be confusing.

Clarify the log message at abnormal module exit to be clear about
what is happening, and exclude the result of strerror() to avoid any
confusion with the real error which should be further up in the logs.