kvs: add date to kvs-primary checkpoint
Problem: It'd be convenient if we knew the date when the
primary namespace was checkpointed.

Solution: When checkpointing the primary namespace, store a json
object with version, rootref, and timestamp, instead of just
the rootref string.  On retrieval, parse appropriately and
retrieve timestamp for output in logs.  Support the original
checkpointing format by checking if the checkpoint object
is a raw blobref string first.

Fixes #3580