python: return JobID from flux.job.submit
Problem: The Flux jobid returned from job.submit() or
SubmitFuture.get_id() is an int, when it would be much more convenient
to return an instance of the JobID class.

Return a JobID object from job.submit_get_id() so that
SubmitFuture.get_id() and job.submit() return JobID instead of int.

Fixes #3614