shell/pmi: support local:: key prefix for PMI2
Problem: Cray MPI tries to share shmem segment info using
PMI2_Info_PutNodeAttr() and PMI2_Info_GetNodeAttr() but the
Flux PMI2 implementation uses the PMI-1 wire protocol, which
only supports global scope.

Add a little hack so that keys prefixed with local:: are stored
with shell-local scope.  Our implementation can tack this on
to keys stored with those functions, and we don't have to change
the protocol.

Keys are stored in the hash with the local:: prefix still attached
to keys so that the namespace is kept separate from global socpe.

Fixes #4141