testsuite: drop kvstest --library option
Problem: The kvstest PMI client includes a --library option
that no longer has an effect on the Flux PMI implementation.

The Flux libpmi.so no longer tries to dlopen another library if
the PMI_LIBRARY environment variable is set, so the kvstest --library
is no longer useful.

Drop this option from test program.