update KVS checkpoint protocol for future changes
Problem: the kvs-checkpoint.get and kvs-checkpoint.put methods provided
by content back end modules operate on (key, blobref) tuples, but we would like
to store other metadata with checkpoints, such as a date, and not have to
change the database schema in every back end whenever the value format changes.

Solution: change the stored value to a json object that contains a blobref, a
date, and a version=1. The back ends just passively shuttle the json object
without interpretation, so they should not need to be changed for updates going
forward.  There is one exception this time: to avoid losing data on system
instance installations when rolling out this change, content-sqlite translates
the old format to a json object with version=0.

Update users of the kvs-checkpoint methods and add tests.

Fixes #4144