job-manager: allow guests to query
Problem: it is useful for guests to know when job submission
or scheduling is disabled, but these admin requests are restricted
to the instance owner.

Instead of gating access in the request table, allow guest access to
job-manager.submit-admin and job-manager.alloc-admin, then within
the request, allow guest to query only.

Update sharness test.