flux-startlog: new utility
Problem: Flux cannot detect that it is being restarted after a broker
crash, so that it can take remedial action or avoid worsening problems.

Add a utility that can manipulate a new eventlog, which operates
somewhat like the last(1) command with wtmp(5).

Rc1 can call startlog --post-start-event to post a start event to eventlog.
Rc3 can call startlog --post-finish-event to post a finish event to eventlog.

If the instance crashes, the finish event will not be written, and the
crash can be detected by two consecutive start statements at the
tail of the log.  'flux startlog --check' tests for this condition.

When run with no options, the past instance run intervals are printed
with durations.