job-exec: rename exec_conf to exec_ctx
Problem: The exec_conf struct in the bulk exec implementation of
the job-exec module will not only be used for configuration state,
so the structure name is not apt.

Rename `struct exec_conf` to `struct exec_ctx` and instances of the
struct from `conf` to `ctx`, but do not change anything else.