job-exec: detect shells that exit before barrier
Problem: If the IMP exits before launching a job shell, or if a job
shell exits before it is connected to Flux and can raise a fatal job
exception, jobs may hang because the exec system does not know that
the IMP/shell exit is due to an abnormal condition (the nonzero exit
status of a shell could be due to a task exiting with nonzero status).

To give the exec system a window into the status of job shell
intialization, support a very simple barrier over a new subprocess
channel, exported via the FLUX_EXEC_PROTOCOL_FD environment variable.
Conformant job shells will use this exec system barrier instead of a
flux_barrier(3) so that the exec system knows when all shells have at
least reached the first execution barrier. If a shell exits before
the first barrier is completed, then the exec system will raise an
exception and terminate the barrier with an error (if any shells have
entered it)

Fixes #4150