testsuite: test job-exec handling of early IMP failure
Problem: No tests ensure that job-exec and the job shell terminate
with an error in a timely fashion when one IMP of a multi-shell job
terminates abnormally.

Add a test imp script which always exits early on broker rank 1
and use this to simulate a case where an IMP exits before invoking
the job shell. Ensure that the job terminates with error as expected.