libflux: add flux_hostmap_lookup(3)
Problem: A function to convert a list of ranks in Idset form to a
list of hostlist in Hostlist form, along with the reverse operation
does not exist in libflux. Instead this functionality is embedded in
the flux-overlay lookup subcommand.

Add a function to libflux with the same functionality as flux-overlay
lookup -- flux_hostmap_lookup(3) takes a Flux handle and string
'targets' to convert from either a Hostlist of hostnames or an Idset
of ranks to the opposite. A string is returned which could be decoded
with idset_decode(3) or hostlist_decode(3) as required, but may also
be useful itself in error or log messages.