shell: remove old-style shell barrier
Problem: The code that falls back to flux_barrier(3) while
asynchronously monitoring the job eventlog for an exception is no
longer used, since the exec system should always set a protocol fd
in FLUX_EXEC_PROTOCOL_FD.

Remove the unused code. If we need to resurrect it someday it will
probably be easy to revert this commit.