libflux/future: set errno in flux_future_wait_for
Problem: When a negative timeout is passed to flux_future_wait_for()
and a future is not properly initialized to be tied to a localized
'now' reactor, flux_future_wait_for() can return -1 without setting
an errno.

Set errno of EINVAL when the above corner case occurs.  Add unit
test.

Fixes #4160