flux-config: add get subcommand
Problem: There is no way to query the broker's configuration object
from a shell script such as rc1.

Add the "flux-config get" subcommand.  If run without arguments, it
dumps the entire config object.  Otherwise, it walks a period-separated
path and dumps the result.

By default, the value is printed as compact, string-encoded JSON, unless
it is a bare string value, in which case the string is printed without
quotes.

If the --type=TYPE option is specified, it is an error if the value does
not match the type, which may be any jansson type (object, array, integer,
string, real, boolean) or one of the following special type names:

  any - value matches any type (the default)
  fsd - value must parse as a FSD, printed as its original string form
  fsd-integer - value must parse as a FSD, printed as integer seconds
  fsd-real - value must parse as a FSD, printed a real seconds

If the requested value is not set, then:
- if no --quiet and no --default, an error is printed and program fails
- if --quiet, nothing is printed and program fails
- if --default, the option argument is printed and program succeeds

Fixes #4161