libutil: add errprintf utility function
Problem: Many flux-core interfaces that use an *error_t to capture
function errors implement an internal errprintf() convenience function
to fill in the error text. This has resulted in a bit of duplicated
code.

Add an errprintf() implementation to libutil which can be used in place
of duplicated code once callers are updated to use flux_error_t.

Also add a set of simple unit tests for the interface.