content-files: call open() with O_TRUNC
Problem: In content-files, most files are created using a hash
of the data.  So in most cases the filenames are unique.  However,
the kvs-checkpoint.put rpc will typically write to the same file
over and over again.  In the event the file already exists, content-files
did not clear the data in a file before writing.  This can lead to
corrupted data if the user wrote fewer bytes to the file than before.

Solution: Call open() with O_TRUNC to truncate the file before writing
new data.  Add tests increasing / decreasing checkpoint data size.