shell: require FLUX_EXEC_PROTOCOL_FD
Problem: The shell fallback barrier code has been removed, but the
shell doesn't check for a failure to obtain FLUX_EXEC_PROTOCOL_FD
from the environment, so when this value is not set an unspecific
"shell barrier failed" error is generated.

Test that shell->protocol_fd is not < 0 in shell_barrier() and report
a more specific error if so.