drop unused project meta files
Problem: automake "foreign mode" was enabled in a previous commit,
but unused GNU project-specific meta files that were required before
that were not cleaned up.

Drop ChangeLog and README which have no content.
Drop the NOTICE link to NOTICE.llns as it is not helping much.
Drop INSTALL as Flux-specific installation instructions are
in the README.md and in readthedocs.

Fixes #4135