broker: add state-machine.get RPC
Problem: There is no way to query the current broker state and
the elapsed time in it.

The only available RPCs are for internal use of the state machine
only, since they use state numbers that are only defined internally.

Add a new RPC, open to guests, that simply returns the current state
name and the elapsed time in it.