flux-uptime: show time in current broker state
Problem: if the broker gets stuck in rc1, cleanup, or rc3,
it may not be evident what is going on without examining logs.

Change flux-uptime so that instead of always showing "up", it shows
the current broker state.  In addition, if the state is not "run",
change the duration to refelect the time the local broker has been
in that state, rather than the overall instance runtime.