build: ensure autogen.sh updates package version
Problem: PACKAGE_VERSION is not updated to match "git describe"
in a development tree when autogen.sh is run.

The previous value is retained in the autom4te.cache directory,
even when configure is regenerated, which seems counterintuitive.

Add the --force option to autoreconf (as called from autogen.sh)
to ensure that nothing is cached from previous builds.