job-list: use flux_msglist for backlog
Problem: handling the journal backlog for pause/unpause with a
zlistx of json objects is slightly more complicated and less
intuitive than placing the response messages in a flux_msglist.

Move to a flux_msglist, which has the following advantages:
- message refcounting is built-in
- deletion possible while iterating

Also rename jsctx->events_journal_backlog to jsctx->backlog, since the
name is long and a bit redundant, given that it is a member of the
journal context struct.