resource: rerank job R based on hostlist attribute when possible
Problem: The mapping of ranks to hosts in the R 'nodelist' key may
not match the actual mapping of ranks to hosts as indicated in the
'hostlist' attribute (which is the definitive mapping). This may occur
when launching a job under Flux and the job shell decides to map ranks
in a different order than the ranks are assigned in the provided R.
For example, currently the job shell sorts ranks by total available
cores such that rank 0 is assigned to the host with the most cores,
thus resulting in a possible different ordering in heterogeneous
allocations.

To ensure that the mapping of hosts to ranks matches the hostlist
attribute, call rlist_rerank() before rlist_remap() in the reosurce
module, when R is provided by a parent instance (i.e. in the case of
a Flux job).

Fixes #4182