etc/rc1: do not cancel jobs under system instance
Problem: By default in etc/rc1, a script is setup to cancel all
jobs when the flux broker is brought down on rank 0.  This should
not be done when running the system instance.

Solution: Check if we are running as the system instance by checking
the configuration of systemd.enable via `flux config get`.  If we
are, do not enable job cancelation.