broker: add assertions for cached attributes
Problem: some broker attributes have to be set with the
IMMUTABLE flag, or mundane functions like flux_get_rank()
will deadlock on an RPC to self.

Make the broker code more serviceable by adding some assertions
that rank, size, and hostlist are in the cache with the IMMUTABLE
flag set, since violating those invariants could result in a
difficult to diagnose (and surprising) hang.

Ensure that the hostlist is set when config bootstrap is selected
and the hosts array is empty or missing.