flux-top: don't re-send already pending RPCs
Problem: if resource or job status queres are slow for some
reason, additional queries are sent whenever there is job
activity, but there is no benefit to having multiple outstanding
queries.

Example: job-list could be slow to respond if it happens to still
be initializing after an instance restart.

Track futures for resource and job-list queries and don't resend
them if still pending.