librlist: add rlist_copy_cores()
Problem: It would be useful to create a copy of a resource set with
only some resource types included, but no such interface exists in
librlist. This would be useful, for instance, to use rlist_verify()
or rlist_diff() only only a subset of resource types.

Add rlist_copy_cores() for now to allow this functionality for cores
only.