resource: restrict resource verification to cores only
Problem: When the system installed libhwloc fails to detect GPUs,
the Flux system administrator may be forced into an untenable situation:
GPUs which are configured in system.R will cause resource module
verification of resources to fail unless the resource.noverify
is set, but this may not help child job instances which also may
fail to discover allocated GPUs.

Until a solution for libhwloc GPU detection is well known, disable
verification of GPU resources completely. This allows GPUs to be
configured in a system instance and propagated via allocation without
needing to worry about propagating the noverify option consistently.