libjob: add format string to flux_job_statetostr()
Problem: flux_job_statetostr() cannot print lower case state names,
so it requires a wrapper in job-list.

Change boolean argument to a format string.  Allow the following
  "s" - lower case short name
  "S" - upper case short name
  "l" - lower case long name
  "L" - upper case long name

Use tables instead of switch statements for the lookup, and use the
table in flux_job_strtostate() as well.

Update users and tests.