testsuite: ensure GPUs are not verified by default
Problem: No tests ensure that GPU resources configured in the system
instance R are not subject to resource module verification.

Add a test to t2315-resource-system.t to ensure that an unlikely set
of GPUs in a configured R doesn't cause the instance to fail or ranks
to be drained.