testsuite: test a couple more scenarios for early shell exit
Problem: No tests in the testsuite exercise job-exec handling of
early job shell exit _after_ the first shell barrier.

Add a couple test cases to the dummy job shell tests which use the
new capability of the dummy shell to exit after a first barrier but
before the second.

This should exercise the special case handling in the job-exec module
for aborting barriers in these scenarios.