flux-restore: add new utility
Problem: there is no way to restore a KVS archive created by flux-dump.

flux-restore reads an archived KVS snapshot created by flux-dump.  It
reads all formats and compression types supported by the libarchive
project.  The snapshot is restored directly to the content store, so
if the --checkpoint option is used, the KVS module does not need to be
loaded.

Otherwise, the blobref of the restored snapshot can be assigned to a
KVS key using --key=NAME.

Value fragmentation is not preserved by flux-restore, so for example
eventlogs that were comitted with one entry per blob are restored as
one blob.

Only regular values and symbolic links are expected to be in the archive.
Directories are only recreated as needed to contain keys, therefore
empty directories are not preserved across flux-dump/flux-restore.

If restoring a checkpoint, the checkpoint timestamp is set to the
newest mtime of a key in the archive.