shell: log hostname for a couple key fatal errors
Problem: Occasionally, when a task fails to start (due to missing
executable for example), or the shell fails to chdir(2), the problem
is node-specific, but the shell logs only the rank of itself or the
task, which is not convenient for end users.

Add the hostname to these two specific error cases.