testsuite: relax match for job exception in t0010-job.py
Problem: The exception tests in python/t0010-job.py require exact
matches, but this has the potential for tests that break with small
changes to the log messages in the future, such as the addition of
a hostname.

Allow the exception "note" to match as a regex instead of an exact
match, unless the note is expected to be empty, in which case a regex
won't work so use an exact match.

Expand the regex for the missing command test, to allow a hostname
to be added in the near future without breaking the test.