flux-resource: differentiate drained vs draining
Problem: The flux-resource drain and status commands do not indicate
ranks in "drain" state that still have jobs with active allocations.
This could cause a node to be shutdown before all running jobs using
the system have completed when targets are drained for maintenance
purposes.

Combine both the resource.status and sched.resource-status RPCs in
`flux resource status` and `flux resource drain` so that the idle
ranks can be identified as "drained" and the ranks still assigned
jobs can be identified as "draining".

In `flux resource status`, keep the "drain" state, which now references
*both* 'drained' and 'draining' ranks. This can be useful for calls
like `flux resource status -s drain -no {nnodes}`, to query the total
number of nodes currently draining or drained.

Fix references to "drain" that need to be either drained or draining
in the testsuite as necessary to avoid breaking tests.

Fixes #4185