flux-dump: add new utility
Problem: there is not an efficient way to generate a backup of the
Flux KVS or generate a portable dump file that could be restored into
a differently configured Flux.

Add flux-dump, which walks the KVS namespace from either a checkpoint
or the current root, and generates a portable archive.  The format and
compression level is determined by the output file extension.  If the
output file is '-', then the file format is tar with no compression.

Because all commits to the KVS are atomic and propagate to the root of
the namespace, any given snapshot starting at the root reflects the entire
KVS state after a the most recent commit.  Thus a dump may be generated on
a live system without content changing during traversal.

Since flux-dump generates the dump by interacting directly with the
content store, a dump may be created when the KVS is not loaded, as
long as the --checkpoint option is supplied so it has a starting blobref.
Otherwise, the starting blobref is obtained by requesting the current
root from the KVS.

Only regular values and symbolic links are dumped to the archive.
Directories are not dumped, so empty directories will not be preserved
across flux-dump/flux-restore.  This is by design so that inactive jobs
can be purged from a system instance without needing to remove now-empty
directories, a process that could be racy on a live system.

Fixes #4202