testsuite: modify expected TypeError assertion for IDset
Problem: A couple IDset class tests in t/python/t0021-idset.py
check for TypeError when performing operations on an int, but it
may actually be useful to allow an int to represent an IDset of
size 1.

Change the type used in these tests to float, so that TypeError will
still be raised when the allowed types for an IDset is relaxed.