broker: block SIGPIPE
Problem: The broker can be killed by SIGPIPE if a component writes
to a closed file descriptor or socket, but it would be better to
handle the error at the call site and continue running, instead of
exiting immediately.

Set up a list of signals to block in broker_handle_signals() and add
SIGPIPE to this list.

Fixes #4183