sharness: preserve --root in test_under_flux
Problem: The sharness --root=PATH option is valuable when running
sharness tests in parallel since it can be used to place the trash
directory for each test in a separate path. However, this option is
not being passed down through test_under_flux(), so the trash directory
reverts to the original location, causing conflicts or one test to
remove the current working directory of another simultaneous test.

If $root is set, propagate --root=$root to the sharness script
re-invocation in test_under_flux().

This allows a flaky script to run multiple times in parallel with
something like

  $ flux start -s 4 \
        flux mini submit --cc=1-100 -o pty --watch --progress \
        sh -c './t2402-job-exec-dummy.t -d -v --root=$TMPDIR'

Since TMPDIR is set to a unique directory per job.