python: add errors=surrogateescape to encode_payload()
Problem: flux.util.encode_payload() fails with unicode errors if
valid utf-8 characters are in the payload and a non-UTF-8 locale
such as LANG=C is set (as is in the case in the testsuite).

Add errors=surrogateescape to the encode() methods used in
encode_payload() to avoid these errors.