testsuite: fix use of --root with system personality tests
Problem: The system personality tests already use the --root=DIR option
to relaunch test instances in a subdirectory of the existing trashdir,
but if --root is also passed on the command line of the test, then
the option gets added twice to the test_under_flux() re-invocation
of the test script. This confuses sharness and causes tests to fail.

Only propagate the current value of $root as --root=$root when it
isn't already set by the system personality.