libkvs: define default checkpoint key
Problem: the checkpoint key name of "kvs-primary" is hardcoded
in several places, and is the only key currently used.

Define KVS_DEFAULT_CHECKPOINT to "kvs-primary", and change
kvs_checkpoint_commit() and kvs_checkpoint_lookup() to use this if the
key parameter is NULL.

Update users to call the functions with a NULL key parameter
instead of hardwiring "kvs-primary".