testsuite: relocate useful sqlite query tool
Problem: testing for content.sqlite dump/restore could benefit from
the sqlite query script developed for job-archive, but that script
is in a job-archive specific directory.

Move script to the t/scripts directory, and change header comment to
mention of job-archive.  There does not appear to be anything specific
to job-archive in the script - it executes generic sql queries on any
sqlite database.