testsuite: avoid use of submit --progress option in testsuite
Problem: The --progress option is used in a couple places in the
testsuite, but this can mangle a user's tty when they are also running
the testsuite under bulksubmit with the --progress option.

Since there is no good reason to use --progress inside of the
testsuite, remove all uses that do not test the option itself.