testsuite: support --root option in fluxometer.lua
Problem: The tests in t/lua/t*.t do not support the --root=PATH option
like sharness, which can be used to ensure the same test run multiple
times on the same node doesn't conflict with itself.

Add support for the --root option in fluxometer.lua so that Lua tests
can be run with the same command line as all other tests.