sharness: pass --root down to sub-sharness during test
Problem: t0000-sharness.t fails when run multiple times per node
because the sub-sharness may run in an original trash direcotry
not one specified to the --root=PATH directory.

Check for $root set in the environment of run_sub_test_lib_test()
and pass it along to the sharness invocation in the function.